Removed firmware from current status
This information is useful but not in the current status page. I plan
on moving this to the about page however I want to avoid an HTTP
request to the OpenSprinkler on intial page load so will probably do it
using AJAX. Regardless, for now it has no purpose in current status.