Backup: Added file based export/import
The primary buttons are still present and perform the same action.
However, new buttons were added to the right side and handle file based
export/import. The buttons are disabled on iOS since no files can be
downloaded or uploaded. Tested on Chrome/Opera on the desktop and
Android devices.